+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Breaking changes are annotated with ☢️, and alpha/beta features with 🐥.
 > `v0.18.2`. This typically means that there was some CI/tooling mishap. Ignore
 > those gaps.
 
-## [Unreleased]
+## [v0.55.0] - 2026-09-09
 
 ### Added
 
@@ -1845,3 +1845,4 @@ make working with lots of sources much easier.
 [v0.53.0]: https://github.com/neilotoole/sq/compare/v0.52.0...v0.53.0
 [v0.54.0]: https://github.com/neilotoole/sq/compare/v0.53.0...v0.54.0
 [v0.54.1]: https://github.com/neilotoole/sq/compare/v0.54.0...v0.54.1
+[v0.55.0]: https://github.com/neilotoole/sq/compare/v0.54.1...v0.55.0